+++ src/parser/inblock/autolink.ts
@@ -1,11 +1,11 @@
 ﻿import { AutolinkParser } from '../inblock';
-import { union } from '../../combinator';
+import { union, build } from '../../combinator';
 import { email } from '../inline';
 import { channel } from './autolink/channel';
 import { hashtag } from './autolink/hashtag';
 
-export const autolink: AutolinkParser = union([
+export const autolink: AutolinkParser = build(() => union([
   email,
   channel,
   hashtag,
-]);
+]));